[tests] refactor: drop tests/pipelines/conftest.py in favor of direct imports
The conftest existed only to serve two Qwen3-Omni CPU test files and
loaded their modules under test via spec_from_file_location with fake
verl_omni packages, on the theory that the package init pulls in GPU
deps. That is not true in practice: vllm and vllm-omni are installed
in local and CI environments, the package imports fine on CPU, and
100+ other tests in this repo already use plain verl_omni imports.

Import the adapter and omni_rl_datasets directly, inline the tiny
processor/tokenizer stubs the parity test needs, and replace the
require_version fixture with a local helper.

AI assistance (OpenCode) was used for this change.
